task 5

+++ 0x0E-structures_typedef/4-new_dog.c
@@ -31,9 +31,9 @@ dog_t *new_dog(char *name, float age, char *owner)
 
 		return (NULL);
 	}
-	for(i = 0; i < strlen(name); i++)
+	for(i = 0; i < (int)strlen(name); i++)
 		doggy->name[i] = name[i];
-	for (j = 0; j < strlen(owner); j++)
+	for (j = 0; j < (int)strlen(owner); j++)
 		doggy->owner[j] = owner[j];
 
 	doggy->age = age;
